refactor: index ページコンポーネントをディレクトリ構造に再編成
_internal/components/ 配下の各コンポーネントを
独立したディレクトリに移動し、スタイルファイルを同梱

- Heading/: Heading.astro + styles.css.ts
- KeyVisual/: KeyVisual.astro + styles.css.ts
- ProfileSection/: ProfileSection.astro + styles.css.ts
- WorkSection/: WorkSection.astro + styles.css.ts
- index.astro: import パスを更新
- index.css.ts: 削除（各コンポーネントに分散）
