+++ src/pages/_internal/Heading/Heading.astro
@@ -1,5 +1,5 @@
 ---
-import { root } from './Heading.css';
+import { root } from './styles.css';
 ---
 
 <h1 class={root}>
