+++ src/pages/_internal/Heading/styles.css.ts
@@ -1,6 +1,6 @@
 import { style } from '@vanilla-extract/css';
 
-import { vars } from '../../../styles/theme.css';
+import { vars } from '../../../foundation/styles/theme.css';
 
 export const root = style({
 	fontSize: '32px',
