+++ src/pages/_internal/KeyVisual/KeyVisual.astro
@@ -7,7 +7,7 @@ import KvPhrase from '../../../assets/kv-phrase.svg';
 import KvPhraseSm from '../../../assets/kv-phrase-sm.svg';
 import Scroll from '../../../assets/scroll.svg';
 
-import { root, main, image, background, phrase, scroll, scrollText, phraseImage, phraseImageSm } from './KeyVisual.css';
+import { root, main, image, background, phrase, scroll, scrollText, phraseImage, phraseImageSm } from './styles.css';
 ---
 
 <div class={root}>
