+++ src/pages/_internal/WorkSection/styles.css.ts
