--- src/pages/_internal/components/WorkSection.astro
@@ -1,30 +1,0 @@
----
-import Heading from './Heading.astro';
-import LinkButton from '../../../components/LinkButton.astro';
-import { WorkCard, type WorkCardProps } from '../../../features/work/components/WorkCard';
-
-import { root, worksList, work, linkButton } from './WorkSection.css';
-
-import { getCollection } from 'astro:content';
-import { collectionToWork, sortWorksInDisplayOrder } from '../../../context/work/services';
-import { processImageForWorkCard } from '../../../features/work/services';
-
-const worksCollectionEntries = await getCollection('works');
-const works = sortWorksInDisplayOrder(worksCollectionEntries.map((work) => collectionToWork(work)));
-const pickedUpWorks = works.filter((work) => work.pickUp);
-const workCardProps: WorkCardProps[] = await Promise.all(pickedUpWorks.map((work) => processImageForWorkCard(work)));
----
-
-<section class={root}>
-	<Heading>works</Heading>
-	<ul class={worksList}>
-		{
-			workCardProps.map((workCardProps) => (
-				<li class={work}>
-					<WorkCard {...workCardProps} />
-				</li>
-			))
-		}
-	</ul>
-	<LinkButton class={linkButton} href="/works">see all</LinkButton>
-</section>
