--- src/pages/index.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const content = style({
-	display: 'flex',
-	flexDirection: 'column',
-	gap: '64px',
-});
